Update miscellaneous-settings.js
Added the ability to read a CSS file, parse it, then look up values by using the selector and the attribute name. Then sets the default color values for the Color Picker. Was added so as the CSS Style Sheet changes for the selectors, there would be no need to update the Color Picker Defaults for color in this file.

Note:
This is using an Ajax call in synchronous mode. (async: false) = deprecated code.
Otherwise the file is not finished reading before setting the Color Picker default colors.
Will update it later when I find a better solution.

~Stevil