Put back the 'portrait' tag like it was.
I had made it so that it didn't save the name in the 'portrait' tag because of a special case where one of the GCA files I was using as a test had an & in the file name. When importing the xml file, it would fail. I put it back to normal as it should be fixed inside of the GGA.