Update ExportToFoundryVTT.vb
This includes a new function to return a concatenated UserNotes and VTT Notes. It includes a plain-texted UserNotes first, then a line break, then VTT Notes. Only non-empty fields are included, and adds no line-break if just one field is included. The new function is now used in place of the previous calls to plain-text UserNotes. Hopefully this successfully allows users to enjoy both fields as intended.