Added experimental actor sheet (unfinished), started on actor sheet AppV2 migration
The experimental actor sheet is a rewrite of the existing actor sheet,
from the ground up. It matches the visual style of Foundry AppV2
applications and attempts to clean up some functionality. The sheet is
currently not enabled for users.

The actor sheet AppV2 migration is a minimal migration aiming to produce
as close as possible to the same sheet as the AppV1 sheet, except using
the AppV2 API instead