Fixed aim applying twice when mulitple attack modes are present
Added case under roll handling for when data-key is present in order to
extract more information from the object than just using data-otf would
produce. This information is used to discriminate against modifiers
which should not be applied.

Also, simplified range penalties because an easy way to do seemed to be
introduced in v12.